make RadarDisplay_airborne subclass of RadarDisplay
